+++ src/main/java/com/chat/app/services/base/UserService.java
@@ -1,17 +1,16 @@
 package com.chat.app.services.base;
 
+import com.chat.app.models.UserDetails;
 import com.chat.app.models.UserModel;
 import com.chat.app.models.specs.NewPasswordSpec;
-import com.chat.app.models.specs.RegisterSpec;
 import com.chat.app.models.specs.UserSpec;
-import org.springframework.security.core.userdetails.UserDetails;
 
 import java.util.List;
 
 public interface UserService {
     List<UserModel> findAll();
 
-    UserModel findById(int id);
+    UserModel findById(long id);
 
     List<UserModel> findByUsernameWithRegex(String username);
 
@@ -19,7 +18,7 @@ public interface UserService {
 
     void delete(long id, UserDetails loggedUser);
 
-    UserModel changeUserInfo(int loggedUser, UserSpec userSpec);
+    UserModel changeUserInfo(long loggedUser, UserSpec userSpec);
 
     UserModel changePassword(NewPasswordSpec changePasswordSpec);
 }
